refactored DeckboxClone to use TS
